Refactor: liblrmd: rename internal header
... from lrmd_alerts_internal.h to lrmd_internal.h, so it can be used for stuff
other than alerts.
