+++ daemons/controld/controld_execd_state.c
@@ -13,7 +13,7 @@
 #include <crm/common/iso8601.h>
 #include <crm/pengine/rules.h>
 #include <crm/pengine/rules_internal.h>
-#include <crm/lrmd_alerts_internal.h>
+#include <crm/lrmd_internal.h>
 
 #include <pacemaker-internal.h>
 #include <pacemaker-controld.h>
