+++ include/crm/Makefile.am
@@ -14,6 +14,6 @@ headerdir=$(pkgincludedir)/crm
 header_HEADERS		= attrd.h cib.h cluster.h compatibility.h crm.h \
 			  lrmd.h msg_xml.h services.h stonith-ng.h
 
-noinst_HEADERS		= lrmd_alerts_internal.h
+noinst_HEADERS		= lrmd_internal.h
 
 SUBDIRS                 = common pengine cib fencing cluster
