+++ lib/lrmd/lrmd_alerts.c
@@ -1,5 +1,7 @@
 /*
- * Copyright 2015-2018 David Vossel <davidvossel@gmail.com>
+ * Copyright 2015-2019 the Pacemaker project contributors
+ *
+ * The version control history for this file may have further details.
  *
  * This source code is licensed under the GNU Lesser General Public License
  * version 2.1 or later (LGPLv2.1+) WITHOUT ANY WARRANTY.
@@ -16,7 +18,7 @@
 #include <crm/common/mainloop.h>
 #include <crm/common/alerts_internal.h>
 #include <crm/common/iso8601_internal.h>
-#include <crm/lrmd_alerts_internal.h>
+#include <crm/lrmd_internal.h>
 
 #include <crm/pengine/status.h>
 #include <crm/cib.h>
